More test re-org
Forgot about the spdx::perf module
